@@ -1,62 +1,103 @@
 <script lang="ts">
+  import type { Snippet } from 'svelte';
   import type { HTMLDialogAttributes } from 'svelte/elements';
-  interface $$restProps extends HTMLDialogAttributes {}
 
-  export let { open = false, className } = $$props as {
-    open: boolean;
+  interface Props extends HTMLDialogAttributes {
+    open?: boolean;
     className?: string;
-  };
 
-  let pane: HTMLDialogElement;
-
-  $: if (pane) {
-    if (open) pane.showModal();
-    else pane.close();
+    main?: Snippet;
+    free?: Snippet;
+    header?: Snippet;
+    footer?: Snippet;
   }
+
+  let {
+    main,
+    free,
+    header,
+    footer,
+    className,
+    open = $bindable(false),
+    ...restProps
+  }: Props = $props();
+
+  let pane: HTMLDialogElement | null = $state(null);
+
+  let preClose: 'open' | 'close' = $state('open');
+
+  $effect(() => {
+    if (pane) {
+      if (open) pane.showModal();
+      else pane.close();
+    }
+  });
 </script>
 
-<!-- svelte-ignore a11y-no-noninteractive-element-interactions -->
 <dialog
   class="Pane"
-  {...$$restProps}
+  class:Pane--preClose={preClose === 'close'}
+  {...restProps}
   bind:this={pane}
-  on:close
-  on:keydown={(e) => {
-    if (e.key === 'Escape') {
+  onkeydown={(e) => {
+    if (e.key === 'Escape' && pane && open) {
       pane.close();
       open = false;
     }
   }}
-  on:click|self={() => {
-    pane.close();
-    open = false;
+  onclick={(e: Event) => {
+    // if the click is outside the pane, close it
+    if (e.target !== e.currentTarget) return;
+    // if the click is on the pane itself, close it
+    if (e.target === pane && pane && open) {
+      // set a delay for 0.5s to allow for the animation to finish
+      preClose = 'close';
+      setTimeout(
+        () => {
+          if (pane && open) {
+            pane.close();
+            open = false;
+            preClose = 'open';
+          }
+        },
+        0.5 * 1000 - 0
+      );
+    }
   }}
 >
-  <!-- svelte-ignore a11y-click-events-have-key-events -->
-  <!-- svelte-ignore a11y-no-static-element-interactions -->
-  <div on:click|stopPropagation>
+  <!-- svelte-ignore a11y_click_events_have_key_events -->
+  <!-- svelte-ignore a11y_no_static_element_interactions -->
+  <div onclick={(e: Event) => e.stopPropagation()}>
     <header>
-      <slot name="header" />
+      {#if header}
+        {@render header()}
+      {/if}
       <button
         class="CrispButton Pane--close"
         data-type="danger"
-        on:click={() => {
-          pane.close();
-          open = false;
+        aria-label="Close"
+        title="Close"
+        type="button"
+        onclick={() => {
+          if (pane && open) {
+            pane.close();
+            open = false;
+          }
         }}
-        data-icon={String.fromCharCode(58829)}
-      />
+        data-icon="close"
+      ></button>
     </header>
-    {#if $$slots.main}
+    {#if main}
       <main>
-        <slot name="main" />
+        {@render main()}
       </main>
-    {:else}
-      <slot name="free" />
     {/if}
-    {#if $$slots.footer}
+    {#if free}
+      {@render free()}
+    {/if}
+    {#if footer}
       <footer>
-        <slot name="footer" />
+        {@render footer()}
       </footer>
     {/if}
   </div>
@@ -65,23 +106,13 @@
 <style lang="scss">
   .Pane {
     --pane-border: 1px solid var(--t-crp-border);
+    --animation: cubic-bezier(0.25, 0.1, 0.25, 1);
     --pane-box-shadow: var(--t-crp-elevation-shadow);
 
-    &__backdrop {
-      top: 0;
-      left: 0;
-      z-index: -1;
-      position: fixed;
-      @include box(100vw, 100dvh);
-    }
-
-    // top: 0;
-    // right: 0;
     border: none;
     margin-top: 0;
     margin-right: 0;
     margin-bottom: 0;
-    // position: fixed;
     margin-left: auto;
     @include box(var(--paneWidth, 600px), 100dvh);
 
@@ -105,7 +136,7 @@
       @include respondAt(600px) {
         @include box(100vw);
       }
-      // @include box();
+
       & > header {
         @include box(100%, auto);
         padding: 16px 24px;
@@ -129,6 +160,14 @@
       }
     }
 
+    &--preClose {
+      animation: slideOut 0.4s var(--animation) forwards !important;
+
+      &::backdrop {
+        animation: fadeOut 0.7s ease-out forwards !important;
+      }
+    }
+
     &::backdrop {
       @include box(100vw, 100vh);
       backdrop-filter: blur(1px);
@@ -136,10 +175,10 @@
     }
 
     &[open] {
-      animation: slideIn 0.2s cubic-bezier(0.87, 0, 0.13, 1);
+      animation: slideIn 0.4s var(--animation);
 
       &::backdrop {
-        animation: fadeIn 0.2s ease-out;
+        animation: fadeIn 0.5s ease-out;
       }
     }
 
@@ -165,12 +204,30 @@
       }
     }
 
+    @keyframes slideOut {
+      from {
+        transform: translateX(0);
+      }
+      to {
+        transform: translateX(100%);
+      }
+    }
+
     @keyframes fadeIn {
       from {
         opacity: 0;
       }
       to {
         opacity: 1;
+      }
+    }
+
+    @keyframes fadeOut {
+      from {
+        opacity: 1;
+      }
+      to {
+        opacity: 0;
       }
     }
   }
