@@ -1,7 +1,12 @@
 <script lang="ts">
   import type { PageData } from './$types';
-  import copyToClipboard from '$utils/CopyToClipboard';
-  export let data: PageData;
+  import { clipboard } from '$utils/CopyToClipboard';
+
+  let {
+    data
+  }: {
+    data: PageData;
+  } = $props();
 </script>
 
 <svelte:head>
@@ -18,13 +23,10 @@
       <div class="Club__header">
         <h2>{club.name}</h2>
         <button
+          data-icon="mail"
           class="CopyButton"
-          data-icon={String.fromCharCode(57688)}
-          on:keydown={async () => await copyToClipboard(club.email)}
-          on:keyup={async () => await copyToClipboard(club.email)}
-          on:keypress={async () => await copyToClipboard(club.email)}
-          on:click={async () => await copyToClipboard(club.email)}
           title="Click to copy email"
+          use:clipboard={{ text: club.email }}
         >
           {club.email}
         </button>
@@ -60,5 +62,5 @@
 </ul>
 
 <style lang="scss">
-  @import '../ClubStyle.scss';
+  @forward '../ClubStyle.scss';
 </style>
