@@ -6,7 +6,11 @@
   import icon_newsletter from '$images/icon_newsletter.png';
   import icon_default from '$images/icon_default.png';
 
-  export let data: PageData;
+  let {
+    data
+  }: {
+    data: PageData;
+  } = $props();
 </script>
 
 <svelte:head>
@@ -45,7 +49,7 @@
           <a href={document.link} target="_blank" rel="noopener noreferrer">
             {#if document.type === 'pdf'}
               <img src={icon_pdf} alt={document.type} />
-            {:else if document.type === 'sheet'}
+            {:else if document.type === 'sheet' || document.type === 'excel'}
               <img src={icon_sheet} alt={document.type} />
             {:else if document.type === 'doc'}
               <img src={icon_doc} alt={document.type} />
@@ -74,13 +78,14 @@
 
       @include respondAt(680px) {
         padding: 0;
-        margin: 40px 0 30px 0;
+        margin: 40px 0 0px 0;
         @include make-flex($just: flex-start, $align: flex-start);
       }
 
       & > h2 {
         font-size: 45px;
         font-weight: 900;
+        line-height: normal;
 
         @include respondAt(500px) {
           font-size: 32px;
@@ -103,7 +108,7 @@
       width: 100%;
       display: grid;
       list-style: none;
-      margin-top: 25px;
+      margin-top: 5px;
       grid-template-columns: repeat(auto-fill, minmax(320px, 1fr));
 
       &--item {
@@ -129,7 +134,7 @@
           }
 
           & > img {
-            width: 60px;
+            width: 50px;
             aspect-ratio: 1/1;
           }
 
