@@ -1,13 +1,13 @@
 <script lang="ts">
-  export let {
+  const {
     width = 27,
     height = 27,
     style
-  } = $$props as {
+  }: {
     width?: number;
     height?: number;
     style?: string;
-  };
+  } = $props();
 </script>
 
 <svg
