@@ -1,14 +1,15 @@
 <script lang="ts">
   import type { IAcademics } from '$types/Amenity.types';
   import MapsSelector from '$utils/MapsSelector';
-  export const data = $$props as IAcademics;
+
+  const data: IAcademics = $props();
 </script>
 
 <div class="AmenityCard">
   <h3>{data.name}</h3>
   <a
     class="CrispButton AmenityCard__bottom--map"
-    data-icon={String.fromCharCode(58715)}
+    data-icon="map"
     href={MapsSelector(data.lat, data.lng)}
     target="_blank"
     rel="noopener noreferrer"
@@ -27,14 +28,9 @@
     background: var(--prof-card-background-color);
     @include make-flex($dir: row, $just: space-between);
 
-    &__bottom {
-      margin-top: auto;
-      @include box($height: fit-content);
-      @include make-flex($just: flex-end, $align: flex-end);
-      &--map {
-        text-decoration: none;
-        background: var(--prof-card-background-color);
-      }
+    &__bottom--map {
+      text-decoration: none;
+      background: var(--prof-card-background-color);
     }
   }
 </style>
